# src/iam/api.yaml
openapi: 3.1.0
info:
  title: Mojaloop IAM provisioning
  version: 1.0.0
  description: >-
    The write side of the platform's authorization, and the only holder of a
    Keto write URL. Services call it when they create a resource; the role
    administration API calls it to read the catalog and to move assignments.

    It carries no gateway route: everything here is reachable only from inside
    the deployment, so the operations declare no security scheme and no
    scoping rather than claiming what nothing enforces. What the platform's
    operators may do with roles and assignments is gated on the way in, by the
    role administration service that fronts this one.

x-authz:
  service: iamProvisioning

servers:
  - url: /

paths:
  /health:
    get:
      operationId: getHealth
      summary: Reports whether the service is running
      security: []
      responses:
        '200':
          description: The service is running

  /state:
    get:
      operationId: getState
      summary: Reports what this process applied, and from what
      description: >-
        The catalog it validated against, the roles and exclusions it read, and
        the record of applying them: how many tuples, which role instances were
        refreshed or retired, and every holding that breaks an exclusion.
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: What this process is working from

  /catalog:
    get:
      operationId: getCatalog
      summary: Lists every service's permissions
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: The composed catalog, one entry per service

  /resource-names:
    get:
      operationId: getResourceNames
      summary: Serves the deployment's vocabulary, verbatim
      description: >-
        Every noun with its label, description, members and source, exactly as
        the deployment declared it, for any consumer to take the fields it
        needs.
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: The vocabulary, one entry per noun

  /keto-namespaces.ts:
    get:
      operationId: getKetoNamespaces
      summary: Serves the composed Keto model
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: The OPL model, one class per service
          content:
            text/plain:
              schema:
                type: string

  /derivation.txt:
    get:
      operationId: getDerivation
      summary: Serves the composed derivation table
      description: >-
        One line per operation stating what the generator concluded, for a
        human to review and diff.
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: The derivation, one section per service
          content:
            text/plain:
              schema:
                type: string

  /access-rules.yml:
    get:
      operationId: getAccessRules
      summary: Serves the composed gateway rules
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: The rules, per service in service order
          content:
            text/plain:
              schema:
                type: string

  /roles:
    get:
      operationId: getRoles
      summary: Lists the deployment's roles and what each leaves open
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: The roles, each with its grants and the resource names an assignment must name

  /resources:
    get:
      operationId: getResources
      summary: Lists the resources that exist, to choose among
      description: >-
        What the registry holds, optionally of one resource name. A resource
        appears here when the service that owns it provisions it, which is
        what lets a role screen offer it by name.
      security: []
      x-authz:
        scopedBy: []
      parameters:
        - name: resourceName
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: The resources that exist

  /subjects/{subjectId}/assignments:
    parameters:
      - name: subjectId
        in: path
        required: true
        description: The identity the assignments belong to
        schema:
          type: string
    get:
      operationId: getSubjectAssignments
      summary: Lists what one identity holds, one entry per role instance
      security: []
      x-authz:
        scopedBy: []
      responses:
        '200':
          description: The role instances this identity is a member of
    post:
      operationId: assignSubjectRole
      summary: Adds an identity to a role
      description: >-
        Takes the resources the role's grants leave open, keyed by resource
        name. Refused when they are not exactly the ones the role takes, when
        one names a resource that does not exist, or when the roles the
        identity would then hold reach both sides of an exclusion.
      security: []
      x-authz:
        scopedBy: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                resources:
                  type: object
                  additionalProperties:
                    type: string
              required: [role]
      responses:
        '200':
          description: The role instance the identity now holds
        '400':
          description: The body is not an assignment
        '422':
          description: The assignment is refused, with the reasons

    delete:
      operationId: unassignSubjectRole
      summary: Removes an identity from a role
      security: []
      x-authz:
        scopedBy: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                resources:
                  type: object
                  additionalProperties:
                    type: string
              required: [role]
      responses:
        '200':
          description: The role instance the identity no longer holds
        '400':
          description: The body is not an assignment
        '422':
          description: The removal is refused, with the reasons

  /provision:
    post:
      operationId: provisionResource
      summary: Records that a resource exists
      description: >-
        A service that creates a resource names only it, keyed by the resource
        name the deployment configured into it. Who holds which role over it
        arrives as ordinary assignments, carrying role names from the same
        configuration.
      security: []
      x-authz:
        scopedBy: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resourceName:
                  type: string
                  minLength: 1
                id:
                  type: string
                  minLength: 1
              required: [resourceName, id]
      responses:
        '200':
          description: The resource is recorded
        '400':
          description: The body does not name a resource
        '422':
          description: The resource name is not one the deployment declares
    delete:
      operationId: deprovisionResource
      summary: Removes a resource and the roles over it
      description: >-
        Answers with the subjects left holding no role at all, so the caller
        can retire an identity that operates nothing else.
      security: []
      x-authz:
        scopedBy: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resourceName:
                  type: string
                  minLength: 1
                id:
                  type: string
                  minLength: 1
              required: [resourceName, id]
      responses:
        '200':
          description: What was removed, and who is left holding nothing
        '400':
          description: The body does not name a resource
        '422':
          description: The resource name is not one the deployment declares